requirements.txt
beautifulsoup4>=4.12.0
docx2txt>=0.8
google-genai>=0.2.0
playwright>=1.46.0
pypdf>=4.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3